+++ src/ffmpegio/__init__.py
@@ -62,7 +62,7 @@ __all__ = ["ffmpeg_info", "get_path", "set_path", "is_ready", "ffmpeg", "ffprobe
     "open", "ffmpegprocess", "FFmpegError", "FilterGraph", "FFConcat"]
 # fmt:on
 
-__version__ = "0.8.4"
+__version__ = "0.8.5"
 
 ffmpeg_info = path.versions
 set_path = path.find